fix for violethold's activation crystals
